+++ maraboupy/Marabou.py
@@ -61,7 +61,7 @@ def read_tf(filename, inputNames=None, outputNames=None, modelType="frozen", sav
     """
     return MarabouNetworkTF(filename, inputNames, outputNames, modelType, savedModelTags)
 
-def read_onnx(filename, inputNames=None, outputNames=None, reindexOutputVars=True):
+def read_onnx(filename, inputNames=None, outputNames=None, reindexOutputVars=False):
     """Constructs a MarabouNetworkONNX object from an ONNX file
 
     Args: