+++ src/input_parsers/NetworkParser.cpp
@@ -58,11 +58,43 @@ void NetworkParser::setUpperBound( Variable var, float value )
 void NetworkParser::addRelu( Variable inputVar, Variable outputVar )
 {
     _reluList.append( new ReluConstraint( inputVar, outputVar ) );
+    setLowerBound( outputVar, 0.0f );
+}
+
+void NetworkParser::addLeakyRelu( Variable inputVar, Variable outputVar, float alpha )
+{
+    _leakyReluList.append( new LeakyReluConstraint( inputVar, outputVar, alpha ) );
 }
 
 void NetworkParser::addSigmoid( Variable inputVar, Variable outputVar )
 {
     _sigmoidList.append( new SigmoidConstraint( inputVar, outputVar ) );
+    setLowerBound( outputVar, 0.0 );
+    setUpperBound( outputVar, 1.0 );
+}
+
+void NetworkParser::addTanh( Variable inputVar, Variable outputVar )
+{
+    // Uses the identity `tanh(x) = 2 * sigmoid(2x) - 1` to implement
+    // it terms of a sigmoid constraint.
+    Variable firstAffine = getNewVariable();
+    Variable sigmoidOutput = getNewVariable();
+
+    Equation e1;
+    e1.addAddend( 2.0, inputVar );
+    e1.addAddend( -1.0, firstAffine );
+    e1.setScalar( 0.0 );
+
+    Equation e2;
+    e2.addAddend( 2.0, sigmoidOutput );
+    e2.addAddend( -1.0, outputVar );
+    e2.setScalar( 1.0 );
+
+    addEquation( e1 );
+    addSigmoid( firstAffine, sigmoidOutput );
+    addEquation( e2 );
+    setLowerBound( outputVar, -1.0 );
+    setUpperBound( outputVar, 1.0 );
 }
 
 void NetworkParser::addMaxConstraint( Variable var, Set<Variable> elements )
@@ -105,40 +137,59 @@ void NetworkParser::getMarabouQuery( InputQuery& query )
 
     for ( ReluConstraint* constraintPtr : _reluList )
     {
-        ReluConstraint constraint = *constraintPtr;
-        ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        DEBUG({
+            ReluConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        });
+        query.addPiecewiseLinearConstraint( constraintPtr );
+    }
+
+    for ( LeakyReluConstraint* constraintPtr : _leakyReluList )
+    {
+        DEBUG({
+            LeakyReluConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        });
         query.addPiecewiseLinearConstraint( constraintPtr );
     }
 
     for ( SigmoidConstraint* constraintPtr : _sigmoidList )
     {
-        SigmoidConstraint constraint = *constraintPtr;
-        ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        DEBUG({
+            SigmoidConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        });
         query.addNonlinearConstraint( constraintPtr );
     }
 
     for ( MaxConstraint* constraintPtr : _maxList )
     {
-        MaxConstraint constraint = *constraintPtr;
-        ASSERT( constraint.getF() < _numVars );
-        for ( [[maybe_unused]] Variable var : constraint.getElements() )
-        {
-            ASSERT ( var < _numVars );
-        }
+        DEBUG({
+            MaxConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getF() < _numVars );
+            for ( [[maybe_unused]] Variable var : constraint.getElements() )
+            {
+                ASSERT ( var < _numVars );
+            }
+        });
         query.addPiecewiseLinearConstraint( constraintPtr );
     }
 
     for ( AbsoluteValueConstraint* constraintPtr : _absList )
     {
-        AbsoluteValueConstraint constraint = *constraintPtr;
-        ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        DEBUG({
+            AbsoluteValueConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        });
         query.addPiecewiseLinearConstraint( constraintPtr );
     }
 
     for ( SignConstraint* constraintPtr : _signList )
     {
-        SignConstraint constraint = *constraintPtr;
-        ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        DEBUG({
+            SignConstraint constraint = *constraintPtr;
+            ASSERT( constraint.getB() < _numVars && constraint.getF() < _numVars );
+        });
         query.addPiecewiseLinearConstraint( constraintPtr );
     }